ajustes flask forms + post
