+++ compilador/app/__init__.py
@@ -1,0 +1,7 @@
+from flask import Flask
+from config import Config
+
+app = Flask(__name__)
+app.config.from_object(Config)
+
+from app import routes
