+++ compilador/app/forms.py
@@ -1,0 +1,7 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, SubmitField, TextAreaField
+from wtforms.validators import DataRequired
+
+class LoginForm(FlaskForm):
+    entrada = TextAreaField('entrada', validators=[DataRequired()])
+    submit = SubmitField('submit')
