+++ compilador/app/templates/index.html
@@ -1,0 +1,5 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <h1>oláá {{ user.username }}</h1>
+{% endblock %}
