+++ compilador/app/templates/processed.html
@@ -1,0 +1,8 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <h1>processed</h1>
+    <p>
+        {{ processo }}
+    </p>
+{% endblock %}
