--- compilador/blueprints/restapi/resources.py
@@ -1,11 +1,0 @@
-from flask import jsonify # serealiza dicionarios 
-from flask_restful import Resource
-
-# class InputCode(code):
-#     def post(code):
-        # ... pegar o input da page e levar ao processamento
-
-
-class ReturnCompiled():
-    def get(self):
-        ... #retorna as infos de compilação
