+++ compilador/config.py
@@ -1,0 +1,4 @@
+import os
+
+class Config(object):
+    SECRET_KEY = os.environ.get('SECRET_KEY') or '1234'
